implementation to tombstone possible in AGR topic

+++ animal-adoptions/src/main/java/com/hiberus/adoptionskafka/services/impl/AdoptionsServiceImpl.java
@@ -1,5 +1,6 @@
 package com.hiberus.adoptionskafka.services.impl;
 
+import com.hiberus.adoptionskafka.exceptions.InstitutionNotFoundException;
 import com.hiberus.adoptionskafka.models.Institution;
 import com.hiberus.adoptionskafka.services.AdoptionsService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.hiberus.adoptionskafka.repositories.InstitutionsRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AdoptionsServiceImpl implements AdoptionsService {
@@ -25,7 +27,12 @@ public class AdoptionsServiceImpl implements AdoptionsService {
     }
 
     @Override
-    public void deleteInstitution(String idInstitution) {
-        institutionsRepository.deleteById(idInstitution);
+    public void deleteInstitution(String idInstitution) throws InstitutionNotFoundException {
+
+        if (institutionsRepository.existsById(idInstitution)) {
+            institutionsRepository.deleteById(idInstitution);
+        } else {
+            throw new InstitutionNotFoundException();
+        }
     }
 }
